+++ DemoApp/CustomMessageContentView.swift
@@ -15,6 +15,13 @@ class CustomMessageContentView: ChatMessageContentView {
             textView?.text = "This message is from a shadow banned user"
         }
         
+        if let translations = content?.translations, let turkishTranslation = translations[.turkish] {
+            textView?.text = turkishTranslation
+            if let timestampLabelText = timestampLabel?.text {
+                timestampLabel?.text = "\(timestampLabelText) - Translated to Turkish"
+            }
+        }
+        
         guard let authorNameLabel = authorNameLabel, authorNameLabel.text != "" else {
             return
         }